move all mappings to one directory RELATIONS
