+++ src/libs/Navigation/linkingConfig/getMatchingBottomTabRouteForState.ts
@@ -2,7 +2,7 @@ import getTopmostCentralPaneRoute from '@libs/Navigation/getTopmostCentralPaneRo
 import type {BottomTabName, NavigationPartialRoute, RootStackParamList, State} from '@libs/Navigation/types';
 import NAVIGATORS from '@src/NAVIGATORS';
 import SCREENS from '@src/SCREENS';
-import {CENTRAL_PANE_TO_TAB_MAPPING} from './TAB_TO_CENTRAL_PANE_MAPPING';
+import {CENTRAL_PANE_TO_TAB_MAPPING} from './RELATIONS/TAB_TO_CENTRAL_PANE_MAPPING';
 
 // Get the route that matches the topmost central pane route in the navigation stack. e.g REPORT -> HOME
 function getMatchingBottomTabRouteForState(state: State<RootStackParamList>, policyID?: string): NavigationPartialRoute<BottomTabName> {
